0.1.3+rev140

+++ src/acpi.cc
@@ -1,7 +1,8 @@
 #include "acpi.hh"
 
+/* ABSOLUTELY WORKS, DUNT EDIT */
 u16 acpi::id() {
-    u16 apic_id;
+    u32 apic_id;
     asm volatile("cpuid" : "=b"(apic_id) : "a"(1));
-    return apic_id >> 24;
+    return (u16)(apic_id >> 24);
 }
